9_vector_databases/8_faiss/requirements.txt
faiss-cpu
numpy
python-dotenv
langchain-openai
langchain-community
langchain-core